# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LowPriceClone",
        "description": "List of endpoints",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/subcategory/store": {
            "post": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Create a new subcategory",
                "operationId": "af4099ce36b5bbd3b08e7a65193050a0",
                "requestBody": {
                    "description": "Subcategory object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categorie_id",
                                    "name"
                                ],
                                "properties": {
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Subcategory Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subcategory created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "SubCategorie added successfully"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Subcategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subcategory/update/{id}": {
            "put": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Mettre à jour une sous catégorie",
                "operationId": "4e0629f6f3d0ea0b3f6222003922dc46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sous catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Objet de la sous catégorie à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categorie_id",
                                    "name"
                                ],
                                "properties": {
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Subcategory Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "subcategory updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Subcategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton d'authentification manquant ou invalide"
                    },
                    "404": {
                        "description": " subcategory  not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subcategory/delete/{id}": {
            "delete": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Supprimer une sous catégorie",
                "operationId": "ba5b56a4dbba43203219348cfab61306",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sous catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "suppression faite"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton d'authentification manquant ou invalide"
                    },
                    "404": {
                        "description": "subcategory  not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subcategory/show/{id}": {
            "get": {
                "tags": [
                    "Subcategories"
                ],
                "summary": "Récupérer les détails d'une sous catégorie",
                "operationId": "425749db597c3a75eae7acb2c3a68c99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sous catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " sous Catégorie récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Subcategory"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton d'authentification manquant ou invalide"
                    },
                    "404": {
                        "description": "SubCategorie non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Subcategory": {
                "title": "Subcategory",
                "required": [
                    "id",
                    "name",
                    "categorie_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "type": "string",
                        "example": "Automobile"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "5"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token should be provided in the 'Authorization' header. Example: Bearer {token}",
                "scheme": "bearer"
            }
        }
    }
}